move examples/books -> examples/http
Since we're adding a new kubernetes gdt plugin, let's move the
examples/books into examples/http to denote they are examples and tests
for the http gdt plugin.
