+++ Makefile
@@ -1,13 +1,12 @@
 VERSION ?= $(shell git describe --tags --always --dirty)
 
-.PHONY: test test-unit test-example ensure-mockery
+.PHONY: test test-unit test-http
 
 test-unit:
 	go test -v ./
 	go test -v ./http
 
-test-example:
-	go test -v ./examples/books/api
-	go test -v ./examples/books/tests/api
+test-http:
+	go test -v ./examples/http/tests/api
 
-test: test-unit test-example
+test: test-unit test-http
