+++ examples/http/api/rest.go
