+++ examples/http/api/suite_test.go.txt
