+++ examples/http/api/types.go
