+++ examples/http/api/types_test.go.txt
