+++ examples/http/tests/api/get_books.yaml
