+++ examples/http/tests/api/put_multiple_books.yaml
