+++ examples/http/tests/api/schemas/get_books.json
